Move react-native related files under 'src' folder
