+++ src/components/splash/AuthLoadingScreen.js
